Now transitions view are inherited from BasicTransitionView
